+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - TT-1727 - Allow inviting collaborator
 - TT-1673 - Handle backfill details after login
 - TT-1748 - Collaboration invitee journey
+- TT-1642 - Handle incomplete companies house details
 
 ### Fixed bugs:
 - TT-1728 - Not ask personal details to individual upgrading to business profile